# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "description": "Opentact API",
        "version": "1.0.0"
    },
    "host": "portal.opentact.org",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
...
        "/api/fs/get_call_handler": {
            "get": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Get call handler",
                "parameters": [
                    {
                        "name": "dnis",
                        "in": "query",
                        "description": "DID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Call handler response"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/fs/join_conference": {
            "get": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Join conference",
                "parameters": [
                    {
                        "name": "dnis",
                        "in": "query",
                        "description": "DID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "ani",
                        "in": "query",
                        "description": "ANI",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "UUID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "conf_name",
                        "in": "query",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "conf_id",
                        "in": "query",
                        "description": "Conference Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/fs/leave_conference": {
            "get": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Leave conference",
                "parameters": [
                    {
                        "name": "dnis",
                        "in": "query",
                        "description": "DID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "ani",
                        "in": "query",
                        "description": "ANI",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "UUID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "conf_name",
                        "in": "query",
                        "description": "Conference name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "conf_id",
                        "in": "query",
                        "description": "Conference Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/fs/agent_join_queue": {
            "get": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Join agent queue",
                "parameters": [
                    {
                        "name": "dnis",
                        "in": "query",
                        "description": "DID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "ani",
                        "in": "query",
                        "description": "ANI",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "UUID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "queue_name",
                        "in": "query",
                        "description": "Queue name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "queue_id",
                        "in": "query",
                        "description": "Queue Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/fs/agent_leave_queue": {
            "get": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Leave agent queue",
                "parameters": [
                    {
                        "name": "dnis",
                        "in": "query",
                        "description": "DID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "ani",
                        "in": "query",
                        "description": "ANI",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "UUID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "queue_name",
                        "in": "query",
                        "description": "Queue name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "queue_id",
                        "in": "query",
                        "description": "Queue Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/dialplan": {
            "post": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Return freeswitch XML",
                "parameters": [
                    {
                        "name": "Caller-ANI",
                        "in": "formData",
                        "description": "Caller ANI",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Caller-Destination-Number",
                        "in": "formData",
                        "description": "Caller Destination Number",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "XML"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "freeswitch"
                ],
                "summary": "Return freeswitch user XML",
                "parameters": [
                    {
                        "name": "user",
                        "in": "formData",
                        "description": "APP user's email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "XML"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}